Keep TypeScript syntax words in skeletons, and split the site-pages suite
The skeleton's reserved set now carries JavaScript's words and
TypeScript's — visibility modifiers, interface, namespace, readonly — so
a widened member or a relabelled declaration reads as different syntax,
not a renamed copy. Split the 637-line site-pages suite into focused
list, create, update, delete, and items files over one shared helpers
module, and pin the rollback test on the loud missing-page lookup.
